requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.158
botocore==1.34.158
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
numpy==1.24.4
openai==1.39.0
pillow==10.4.0
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic-core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
Telethon==1.36.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==1.26.19
